+++ packages/core-cairo/src/erc1155.ts
@@ -11,6 +11,7 @@ import { printContract } from './print';
 import { addSRC5Component } from './common-components';
 import { externalTrait } from './external-trait';
 import { toByteArray } from './utils/convert-strings';
+import { RoyaltyInfoOptions, setRoyaltyInfo, defaults as royaltyInfoDefaults } from './set-royalty-info';
 
 export const defaults: Required<ERC1155Options> = {
   name: 'MyToken',
@@ -19,6 +20,7 @@ export const defaults: Required<ERC1155Options> = {
   pausable: false,
   mintable: false,
   updatableUri: true,
+  royaltyInfo: royaltyInfoDefaults,
   access: commonDefaults.access,
   upgradeable: commonDefaults.upgradeable,
   info: commonDefaults.info
@@ -35,6 +37,7 @@ export interface ERC1155Options extends CommonContractOptions {
   pausable?: boolean;
   mintable?: boolean;
   updatableUri?: boolean;
+  royaltyInfo?: RoyaltyInfoOptions;
 }
 
 function withDefaults(opts: ERC1155Options): Required<ERC1155Options> {
@@ -45,11 +48,12 @@ function withDefaults(opts: ERC1155Options): Required<ERC1155Options> {
     pausable: opts.pausable ?? defaults.pausable,
     mintable: opts.mintable ?? defaults.mintable,
     updatableUri: opts.updatableUri ?? defaults.updatableUri,
+    royaltyInfo: opts.royaltyInfo ?? defaults.royaltyInfo,
   };
 }
 
 export function isAccessControlRequired(opts: Partial<ERC1155Options>): boolean {
-  return opts.mintable === true || opts.pausable === true || opts.updatableUri !== false || opts.upgradeable === true;
+  return opts.mintable === true || opts.pausable === true || opts.updatableUri !== false || opts.upgradeable === true || opts.royaltyInfo?.enabled === true;
 }
 
 export function buildERC1155(opts: ERC1155Options): Contract {
@@ -76,11 +80,12 @@ export function buildERC1155(opts: ERC1155Options): Contract {
     addSetBaseUri(c, allOpts.access);
   }
 
-  addHooks(c, allOpts);
-
   setAccessControl(c, allOpts.access);
   setUpgradeable(c, allOpts.upgradeable, allOpts.access);
   setInfo(c, allOpts.info);
+  setRoyaltyInfo(c, allOpts.royaltyInfo, allOpts.access);
+  
+  addHooks(c, allOpts);
 
   return c;
 }
@@ -95,7 +100,7 @@ function addHooks(c: ContractBuilder, allOpts: Required<ERC1155Options>) {
       priority: 1,
     };
     c.addImplementedTrait(hooksTrait);
-    c.addStandaloneImport('starknet::ContractAddress');
+    c.addUseClause('starknet', 'ContractAddress');
 
     c.addFunction(hooksTrait, {
       name: 'before_update',
@@ -112,7 +117,7 @@ function addHooks(c: ContractBuilder, allOpts: Required<ERC1155Options>) {
       ],
     });
   } else {
-    c.addStandaloneImport('openzeppelin::token::erc1155::ERC1155HooksEmptyImpl');
+    c.addUseClause('openzeppelin::token::erc1155', 'ERC1155HooksEmptyImpl');
   }
 }
 
@@ -136,15 +141,15 @@ function addBase(c: ContractBuilder, baseUri: string) {
 }
 
 function addBurnable(c: ContractBuilder) {
-  c.addStandaloneImport('starknet::ContractAddress');
-  c.addStandaloneImport('starknet::get_caller_address');
+  c.addUseClause('starknet', 'ContractAddress');
+  c.addUseClause('starknet', 'get_caller_address');
   c.addFunction(externalTrait, functions.burn);
   c.addFunction(externalTrait, functions.batch_burn);
   c.addFunction(externalTrait, functions.batchBurn);
 }
 
 function addMintable(c: ContractBuilder, access: Access) {
-  c.addStandaloneImport('starknet::ContractAddress');
+  c.addUseClause('starknet', 'ContractAddress');
   requireAccessControl(c, externalTrait, functions.mint, access, 'MINTER', 'minter');
   requireAccessControl(c, externalTrait, functions.batch_mint, access, 'MINTER', 'minter');
 
@@ -170,11 +175,11 @@ const components = defineComponents( {
       name: 'ERC1155Event',
       type: 'ERC1155Component::Event',
     },
-    impls: [],
-    internalImpl: {
+    impls: [{
       name: 'ERC1155InternalImpl',
+      embed: false,
       value: 'ERC1155Component::InternalImpl<ContractState>',
-    },
+    }],
   },
 });